Adding support for fs.datasync.
This isn't advertised in the Node docs, but it's a standard POSIX function.

It's also fairly reasonable: Most of the time, you probably don't care about syncing mtime/atime/etc, which could potentially be expensive if any cloud storage backends allow for storing that metadata.

I map it onto sync in the default implementation.

We now pass the fsync test.