fs-write-buffer passes, fs-write-file-buffer fails.
It fails because it tries to decode the data in the buffer as a UTF-8 string, and it encounters data that is not a valid UTF-8 character.

Which makes sense; it's *binary* data.

I may have to carry forward my plan to efficiently pack the binary buffer into a UTF-16 JavaScript string as data...