Removing an inapplicable test.
The test is tightly coupled with the implementation. Node's implementation of readFileSync is implemented using its synchronous API -- much like how the default filesystem implementation implements readFile using the other API functions.

It's possible that a filesystem will implement a custom version of this function that does not call closeSync.