Make isAvailable() a static function.
Basically, we want to know if the current browser has the capability to use the file system in general, so it makes more sense this way.

I also changed the tests to check if a file system is available before adding it for testing.

I was prompted to do this because Opera does not support IndexedDB.