FIXED THE DAMN BUG.
document.write is *NOT* innocuous in Firefox/Chrome/Safari, so now we conditionally inject VBScript.

We trigger this bug in Firefox:
https://bugzilla.mozilla.org/show_bug.cgi?id=881487

...and in Chrome/Safari, something completely bizarre happens that causes some global state to vanish (such as our setImmediate polyfill's event listeners), which borks all of our tests.

The worst part: There's no way for the browser to tell you what part of your script causes something like a full page reload, so I had to figure this out using trial-and-error.

Bah.