Fix massive IE<10 speed regression on XHR.
Before BFS, in Doppio, we did the following in IE<10:
* (in VBScript) VBArray => string, each character is a byte.
* Use string as canonical representation of data.

In BFS, we did the following:
* (in VBScript) VBArray => JavaScript array of bytes
* Use this array as canonical representation of data.

Apparently, it's way faster to do the following:
* (in VBScript) VBArray => string, each character is a byte
* (in JavaScript) string => array of bytes.
* Use the array as canonical representation of data.

I believe VBScript was calling JavaScript's Array.push on each byte, which caused a massive slowdown compared to this string-based approach.

Anyway, I'm glad we no longer crash IE9 with a <6MB file. Instead, it takes 5s total to internalize a 5.3MB file, which I can probably improve further at some point.