Emscripten FS frontend for BrowserFS.
Currently works if you force FS.ignorePermissions in Emscripten to `false`; that's due to a BFS bug with permissions that I'm going to look into, report, and fix.

There's currently a bug in Emscripten that prevents it from working for Me and My Shadow (see https://github.com/kripken/emscripten/issues/1893).