Greatly expanding our mode test to cover other invariants.
Fixing a bug in the KVFS that caused us to not set the execute bit on the
root directory.

There are currently other bugs that need to be worked out regarding
appropriate errors associated with writing to RO file systems, hence the
writeFile test is commented out.

Also, this test currently fails. I plan to address the remaining issues
with it soon.

This test is very important, as legacy programs often stat files and check
their modes prior to taking an action.