BrowserFS v0.3.6: Bugfix for EmscriptenFS.
The EmscriptenFS now invokes `utimesSync` whenever it writes a file, but it doesn't really require the value to be set. Many BFS backends do not support `utimesSync`, causing a failure whenever Emscripten wrote a file to BFS.

Thus, I modified the BFS backend to *ignore* `ENOTSUP` errors that occur with `setattr`, which fixes the issue.