Normalize `null` options correctly
Fixes #227.

The first version of Node where FS functions accepted an options object
was version 0.10. Between Node 0.10 and Node 6 inclusive, all falsy
values (including the empty string) and functions are replaced with the
default options. Only truthy values which are not strings, objects or
functions throw a `TypeError`.

Node 7 introduced a breaking change: only `null`, `undefined` and
functions get replaced with the default options, and any values which
are not strings or objects throw a `TypeError`.
https://github.com/nodejs/node/pull/7165/files#diff-9a205ef7ee967ee32efee02e58b3482dR39

It is not clear which version of the Node FS API is emulated by
BrowserFS, so I have elected to implement the Node 7 behaviour.